Move server code out of server __init__.py
