+++ binsync/__main__.py
@@ -143,7 +143,7 @@ def main():
             from binsync.interface_overrides.ghidra import start_ghidra_remote_ui
             start_ghidra_remote_ui()
         elif args.server == "headless":
-            from binsync.extras.server import start_server
+            from binsync.extras.server.server import start_server
             start_server()
         else:
             exit(1)
